pyproject.toml
[tool.poetry]
name = "e-paper-calendar"
version = "0.1.0"
description = ""
authors = ["RyoheiTomiyama <tennisnowboarderyou@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^2.13.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.4"
python-dotenv = "^0.18.0"
numpy = "^1.21.1"
Pillow = "^8.3.1"
python-dateutil = "^2.8.2"
pytz = "^2021.1"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
